refactor: extract review modal form
