+++ src/components/occurrence/review-modal/form/index.ts
@@ -1,0 +1,3 @@
+import Form from './Form';
+
+export default Form;
